move elementwise function to funcs directory
